BACKEND REVISI & DOCUMENTATION DONE

+++ Backend/src/middleware/deleteImg.js
@@ -7,21 +7,25 @@ module.exports = {
 		const id = req.params.id;
 
 		const data = await userModel.selectUserId(id);
-		if (data.rows[0].photo) {
-			const img = data.rows[0].photo;
-			if (img !== "default.png") {
-				fs.unlink(`./assets/profile_pic/${img}`, (err) => {
-					if (err) {
-						res.json({
-							message: "delete failed",
-							error: err,
-						});
-					}
-				});
+		if(data) {
+			if (data.rows[0].photo) {
+				const img = data.rows[0].photo;
+				if (img !== "default.png") {
+					fs.unlink(`./assets/profile_pic/${img}`, (err) => {
+						if (err) {
+							res.json({
+								message: "delete failed",
+								error: err,
+							});
+						}
+					});
+				}
+				next();
+			} else {
+				res.json("There is no profile picture");
 			}
-			next();
-		} else {
-			res.json("There is no profile picture");
+		}else{
+			res.json("Airline ID is not found");
 		}
 	},
 
@@ -29,21 +33,25 @@ module.exports = {
 		const id = req.params.id;
 
 		const data = await airlineModel.getDetailAirline(id);
-		if (data.rows[0].logo) {
-			const img = data.rows[0].logo;
-			if (img !== "default.png") {
-				fs.unlink(`./assets/airline/${img}`, (err) => {
-					if (err) {
-						res.json({
-							message: "delete failed",
-							error: err,
-						});
-					}
-				});
+		if(data.rowCount !== 0){
+			if (data.rows[0].logo) {
+				const img = data.rows[0].logo;
+				if (img !== "default.png") {
+					fs.unlink(`./assets/airline/${img}`, (err) => {
+						if (err) {
+							res.json({
+								message: "delete failed",
+								error: err,
+							});
+						}
+					});
+				}
+				next();
+			} else {
+				res.json("There is no logo");
 			}
-			next();
-		} else {
-			res.json("There is no logo");
+		}else{
+			res.json("Airline ID is not found");
 		}
 	},
 };
